refactor: move train flow to flows
